pavex

Module blueprint

Define the structure of your application using a Blueprint.

Check out the “Project structure” section of the Pavex guide for more details on the role of Blueprint in Pavex applications.

Modules

reflection

Metadata used by Pavex’s CLI to analyze your request handlers, constructors, error handlers, error observers (e.g. their input parameters, their return type, where they are defined, etc.), etc.

Macros

from

Capture a list of sources to be checked by Pavex for components.

Structs

Config

The input type for Blueprint::config.

Constructor

The input type for Blueprint::constructor.

ErrorHandler

The input type for Blueprint::error_handler.

ErrorObserver

The input type for Blueprint::error_observer.

Fallback

The input type for Blueprint::fallback.

Import

The input type for Blueprint::import and Blueprint::routes.

PostProcessingMiddleware

The input type for Blueprint::post_process.

PreProcessingMiddleware

The input type for Blueprint::pre_process.

Prebuilt

The input type for Blueprint::prebuilt.

RegisteredConfig

A configuration type registered via Blueprint::config.

RegisteredConstructor

A constructor registered via Blueprint::constructor.

RegisteredErrorHandler

The type returned by Blueprint::error_handler.

RegisteredErrorObserver

The type returned by Blueprint::error_observer.

RegisteredFallback

The type returned by Blueprint::fallback.

RegisteredImport

The type returned by Blueprint::import.

RegisteredPostProcessingMiddleware

The type returned by Blueprint::post_process.

RegisteredPreProcessingMiddleware

The type returned by Blueprint::pre_process.

RegisteredPrebuilt

The type returned by Blueprint::prebuilt.

RegisteredRoute

The type returned by Blueprint::route.

RegisteredRoutes

The type returned by Blueprint::routes.

RegisteredWrappingMiddleware

The type returned by Blueprint::wrap.

Route

The input type for Blueprint::route.

RoutingModifiers

The type returned by Blueprint::prefix and Blueprint::domain.

WrappingMiddleware

The input type for Blueprint::wrap.

Enums

CloningPolicy

Determines whether Pavex is allowed to clone a type.

Lifecycle

How many times should a constructor be invoked?

Lint

Common mistakes and antipatterns that Pavex tries to catch when analysing your Blueprint.