pavex::cookie

Module errors

Errors that can occur when working with cookies.

Structs

CryptoError

An error that occurred while decrypting or verifying an incoming request cookie.

DecodingError

An error that occurred while decoding a percent-encoded cookie name or value.

EmptyNameError

An error that occurs when parsing a fragment of a Cookie header value that contains an empty name (e.g. =value).

InjectResponseCookiesError

The error type returned by inject_response_cookies.

MissingPairError

An error that occurs when parsing a fragment of a Cookie header value that doesn’t contain a name-value separator (=).

ShortKeyError

The key generation algorithm requires more bytes than what was provided.

Enums

ExtractRequestCookiesError

The error type returned by extract_request_cookies.

KeyError

The error returned by crate::Key::try_from() when trying to create a Key from raw bytes.

ParseError

The error returned by RequestCookies::parse_header().

ProcessIncomingError

The error returned by Processor::process_incoming.

Constants

EXTRACT_REQUEST_COOKIES_ERROR_INTO_RESPONSE

A handle to add ExtractRequestCookiesError::into_response() as an error handler to your Pavex application.

INJECT_RESPONSE_COOKIES_ERROR_INTO_RESPONSE

A handle to add InjectResponseCookiesError::into_response() as an error handler to your Pavex application.