pavex::http

Struct StatusCode

pub struct StatusCode(/* private fields */);

An HTTP status code (status-code in RFC 9110 et al.).

Constants are provided for known status codes, including those in the IANA HTTP Status Code Registry.

Status code values in the range 100-999 (inclusive) are supported by this type. Values in the range 100-599 are semantically classified by the most significant digit. See StatusCode::is_success, etc. Values above 599 are unclassified but allowed for legacy compatibility, though their use is discouraged. Applications may interpret such values as protocol errors.

Examples

use http::StatusCode;

assert_eq!(StatusCode::from_u16(200).unwrap(), StatusCode::OK);
assert_eq!(StatusCode::NOT_FOUND.as_u16(), 404);
assert!(StatusCode::OK.is_success());

Implementations

impl StatusCode

pub fn from_u16(src: u16) -> Result<StatusCode, InvalidStatusCode>

Converts a u16 to a status code.

The function validates the correctness of the supplied u16. It must be greater or equal to 100 and less than 1000.

Example

use http::StatusCode;

let ok = StatusCode::from_u16(200).unwrap();
assert_eq!(ok, StatusCode::OK);

let err = StatusCode::from_u16(99);
assert!(err.is_err());

pub fn from_bytes(src: &[u8]) -> Result<StatusCode, InvalidStatusCode>

Converts a &[u8] to a status code.

pub const fn as_u16(&self) -> u16

Returns the u16 corresponding to this StatusCode.

Note

This is the same as the From<StatusCode> implementation, but included as an inherent method because that implementation doesn’t appear in rustdocs, as well as a way to force the type instead of relying on inference.

Example

let status = http::StatusCode::OK;
assert_eq!(status.as_u16(), 200);

pub fn as_str(&self) -> &str

Returns a &str representation of the StatusCode

The return value only includes a numerical representation of the status code. The canonical reason is not included.

Example

let status = http::StatusCode::OK;
assert_eq!(status.as_str(), "200");

pub fn canonical_reason(&self) -> Option<&'static str>

Get the standardised reason-phrase for this status code.

This is mostly here for servers writing responses, but could potentially have application at other times.

The reason phrase is defined as being exclusively for human readers. You should avoid deriving any meaning from it at all costs.

Bear in mind also that in HTTP/2.0 and HTTP/3.0 the reason phrase is abolished from transmission, and so this canonical reason phrase really is the only reason phrase you’ll find.

Example

let status = http::StatusCode::OK;
assert_eq!(status.canonical_reason(), Some("OK"));

pub fn is_informational(&self) -> bool

Check if status is within 100-199.

pub fn is_success(&self) -> bool

Check if status is within 200-299.

pub fn is_redirection(&self) -> bool

Check if status is within 300-399.

pub fn is_client_error(&self) -> bool

Check if status is within 400-499.

pub fn is_server_error(&self) -> bool

Check if status is within 500-599.

impl StatusCode

pub const CONTINUE: StatusCode

100 Continue [RFC9110, Section 15.2.1]

pub const SWITCHING_PROTOCOLS: StatusCode

101 Switching Protocols [RFC9110, Section 15.2.2]

pub const PROCESSING: StatusCode

102 Processing [RFC2518, Section 10.1]

pub const OK: StatusCode

200 OK [RFC9110, Section 15.3.1]

pub const CREATED: StatusCode

201 Created [RFC9110, Section 15.3.2]

pub const ACCEPTED: StatusCode

202 Accepted [RFC9110, Section 15.3.3]

pub const NON_AUTHORITATIVE_INFORMATION: StatusCode

203 Non-Authoritative Information [RFC9110, Section 15.3.4]

pub const NO_CONTENT: StatusCode

204 No Content [RFC9110, Section 15.3.5]

pub const RESET_CONTENT: StatusCode

205 Reset Content [RFC9110, Section 15.3.6]

pub const PARTIAL_CONTENT: StatusCode

206 Partial Content [RFC9110, Section 15.3.7]

pub const MULTI_STATUS: StatusCode

207 Multi-Status [RFC4918, Section 11.1]

pub const ALREADY_REPORTED: StatusCode

208 Already Reported [RFC5842, Section 7.1]

pub const IM_USED: StatusCode

226 IM Used [RFC3229, Section 10.4.1]

pub const MULTIPLE_CHOICES: StatusCode

300 Multiple Choices [RFC9110, Section 15.4.1]

pub const MOVED_PERMANENTLY: StatusCode

301 Moved Permanently [RFC9110, Section 15.4.2]

pub const FOUND: StatusCode

302 Found [RFC9110, Section 15.4.3]

pub const SEE_OTHER: StatusCode

303 See Other [RFC9110, Section 15.4.4]

pub const NOT_MODIFIED: StatusCode

304 Not Modified [RFC9110, Section 15.4.5]

pub const USE_PROXY: StatusCode

305 Use Proxy [RFC9110, Section 15.4.6]

pub const TEMPORARY_REDIRECT: StatusCode

307 Temporary Redirect [RFC9110, Section 15.4.7]

pub const PERMANENT_REDIRECT: StatusCode

308 Permanent Redirect [RFC9110, Section 15.4.8]

pub const BAD_REQUEST: StatusCode

400 Bad Request [RFC9110, Section 15.5.1]

pub const UNAUTHORIZED: StatusCode

401 Unauthorized [RFC9110, Section 15.5.2]

pub const PAYMENT_REQUIRED: StatusCode

402 Payment Required [RFC9110, Section 15.5.3]

pub const FORBIDDEN: StatusCode

403 Forbidden [RFC9110, Section 15.5.4]

pub const NOT_FOUND: StatusCode

404 Not Found [RFC9110, Section 15.5.5]

pub const METHOD_NOT_ALLOWED: StatusCode

405 Method Not Allowed [RFC9110, Section 15.5.6]

pub const NOT_ACCEPTABLE: StatusCode

406 Not Acceptable [RFC9110, Section 15.5.7]

pub const PROXY_AUTHENTICATION_REQUIRED: StatusCode

407 Proxy Authentication Required [RFC9110, Section 15.5.8]

pub const REQUEST_TIMEOUT: StatusCode

408 Request Timeout [RFC9110, Section 15.5.9]

pub const CONFLICT: StatusCode

409 Conflict [RFC9110, Section 15.5.10]

pub const GONE: StatusCode

410 Gone [RFC9110, Section 15.5.11]

pub const LENGTH_REQUIRED: StatusCode

411 Length Required [RFC9110, Section 15.5.12]

pub const PRECONDITION_FAILED: StatusCode

412 Precondition Failed [RFC9110, Section 15.5.13]

pub const PAYLOAD_TOO_LARGE: StatusCode

413 Payload Too Large [RFC9110, Section 15.5.14]

pub const URI_TOO_LONG: StatusCode

414 URI Too Long [RFC9110, Section 15.5.15]

pub const UNSUPPORTED_MEDIA_TYPE: StatusCode

415 Unsupported Media Type [RFC9110, Section 15.5.16]

pub const RANGE_NOT_SATISFIABLE: StatusCode

416 Range Not Satisfiable [RFC9110, Section 15.5.17]

pub const EXPECTATION_FAILED: StatusCode

417 Expectation Failed [RFC9110, Section 15.5.18]

pub const IM_A_TEAPOT: StatusCode

418 I’m a teapot [curiously not registered by IANA but RFC2324, Section 2.3.2]

pub const MISDIRECTED_REQUEST: StatusCode

421 Misdirected Request [RFC9110, Section 15.5.20]

pub const UNPROCESSABLE_ENTITY: StatusCode

422 Unprocessable Entity [RFC9110, Section 15.5.21]

pub const LOCKED: StatusCode

423 Locked [RFC4918, Section 11.3]

pub const FAILED_DEPENDENCY: StatusCode

424 Failed Dependency [RFC4918, Section 11.4]

pub const TOO_EARLY: StatusCode

425 Too early [RFC8470, Section 5.2]

pub const UPGRADE_REQUIRED: StatusCode

426 Upgrade Required [RFC9110, Section 15.5.22]

pub const PRECONDITION_REQUIRED: StatusCode

428 Precondition Required [RFC6585, Section 3]

pub const TOO_MANY_REQUESTS: StatusCode

429 Too Many Requests [RFC6585, Section 4]

pub const REQUEST_HEADER_FIELDS_TOO_LARGE: StatusCode

431 Request Header Fields Too Large [RFC6585, Section 5]

pub const UNAVAILABLE_FOR_LEGAL_REASONS: StatusCode

451 Unavailable For Legal Reasons [RFC7725, Section 3]

pub const INTERNAL_SERVER_ERROR: StatusCode

500 Internal Server Error [RFC9110, Section 15.6.1]

pub const NOT_IMPLEMENTED: StatusCode

501 Not Implemented [RFC9110, Section 15.6.2]

pub const BAD_GATEWAY: StatusCode

502 Bad Gateway [RFC9110, Section 15.6.3]

pub const SERVICE_UNAVAILABLE: StatusCode

503 Service Unavailable [RFC9110, Section 15.6.4]

pub const GATEWAY_TIMEOUT: StatusCode

504 Gateway Timeout [RFC9110, Section 15.6.5]

pub const HTTP_VERSION_NOT_SUPPORTED: StatusCode

505 HTTP Version Not Supported [RFC9110, Section 15.6.6]

pub const VARIANT_ALSO_NEGOTIATES: StatusCode

506 Variant Also Negotiates [RFC2295, Section 8.1]

pub const INSUFFICIENT_STORAGE: StatusCode

507 Insufficient Storage [RFC4918, Section 11.5]

pub const LOOP_DETECTED: StatusCode

508 Loop Detected [RFC5842, Section 7.2]

pub const NOT_EXTENDED: StatusCode

510 Not Extended [RFC2774, Section 7]

pub const NETWORK_AUTHENTICATION_REQUIRED: StatusCode

511 Network Authentication Required [RFC6585, Section 6]