pavex::request

Module body

Extract data from the body of incoming requests.

Check the relevant section of the guide for a thorough introduction to Pavex’s body extractors.

Modules

errors

Errors that can occur while extracting information from the request body.

Structs

BufferedBody

Buffer the entire body of an incoming request in memory.

JsonBody

Parse the body of an incoming request as JSON.

RawIncomingBody

The raw body of the incoming HTTP request.

UrlEncodedBody

Parse a URL-encoded request body, such as a web form.

Enums

BodySizeLimit

An upper limit on the size of incoming request bodies.