pavex::request::body

Module errors

Errors that can occur while extracting information from the request body.

Structs

JsonContentTypeMismatch

The Content-Type header not set to application/json, or another application/*+json MIME type.

JsonDeserializationError

Something went wrong when deserializing the request body into the specified type.

MissingJsonContentType

The Content-Type header is missing, while we expected it to be set to application/json, or another application/*+json MIME type.

MissingUrlEncodedContentType

The Content-Type header is missing, while we expected it to be set to application/x-www-form-urlencoded.

SizeLimitExceeded

The request body is larger than the maximum size limit enforced by this server.

UnexpectedBufferError

Something went wrong while reading the request body, but we don’t know what specifically.

UrlEncodedBodyDeserializationError

Something went wrong when deserializing the request body into the specified type.

UrlEncodedContentTypeMismatch

The Content-Type header not set to application/x-www-form-urlencoded.

Enums

ExtractBufferedBodyError

The error returned by BufferedBody::extract when the extraction fails.

ExtractJsonBodyError

The error returned by JsonBody::extract when the extraction fails.

ExtractUrlEncodedBodyError

The error returned by UrlEncodedBody::extract when the extraction fails.

Constants

EXTRACT_BUFFERED_BODY_ERROR_INTO_RESPONSE

A handle to add ExtractBufferedBodyError::into_response() as an error handler to your Pavex application.

EXTRACT_JSON_BODY_ERROR_INTO_RESPONSE

A handle to add ExtractJsonBodyError::into_response() as an error handler to your Pavex application.

EXTRACT_URL_ENCODED_BODY_ERROR_INTO_RESPONSE

A handle to add ExtractUrlEncodedBodyError::into_response() as an error handler to your Pavex application.