pavex::request::path

Module errors

Errors that can happen when extracting path parameters.

Structs

DecodeError

The error returned by EncodedParamValue::decode when the percent-decoded path parameter is not a valid UTF8 string.

InvalidUtf8InPathParam

One of the percent-decoded path parameters is not a valid UTF8 string.

PathDeserializationError

Something went wrong when trying to deserialize the percent-decoded URL parameters into the target type you specified—T in PathParams<T>.

Enums

ErrorKind

The kinds of errors that can happen when deserializing into a PathParams.

ExtractPathParamsError

The error returned by PathParams::extract when the extraction fails.

Constants

EXTRACT_PATH_PARAMS_ERROR_INTO_RESPONSE

A handle to add ExtractPathParamsError::into_response() as an error handler to your Pavex application.