pavex::request::query

Module errors

Errors that can happen when extracting query parameters.

Structs

QueryDeserializationError

Something went wrong when trying to deserialize the percent-decoded query parameters into the target type you specified—T in QueryParams<T>.

Enums

ExtractQueryParamsError

The error returned by QueryParams::extract when the extraction fails.

Constants

EXTRACT_QUERY_PARAMS_ERROR_INTO_RESPONSE

A handle to add ExtractQueryParamsError::into_response() as an error handler to your Pavex application.