pavex

Module time

Utilities to work with dates, timestamps and datetimes.

It’s a re-export of the jiff@0.2 crate.

Modules

_documentation

Longer form documentation for Jiff.

civil

Facilities for dealing with inexact dates and times.

fmt

Configurable support for printing and parsing datetimes and durations.

tz

Routines for interacting with time zones and the zoneinfo database.

Structs

Error

An error that can occur in this crate.

SignedDuration

A signed duration of time represented as a 96-bit integer of nanoseconds.

SignedDurationRound

Options for SignedDuration::round.

Span

A span of time represented via a mixture of calendar and clock units.

SpanArithmetic

Options for Span::checked_add and Span::checked_sub.

SpanCompare

Options for Span::compare.

SpanFieldwise

A wrapper for Span that implements the Hash, Eq and PartialEq traits.

SpanRelativeTo

A relative datetime for use with Span APIs.

SpanRound

Options for Span::round.

SpanTotal

Options for Span::total.

Timestamp

An instant in time represented as the number of nanoseconds since the Unix epoch.

TimestampArithmetic

Options for Timestamp::checked_add and Timestamp::checked_sub.

TimestampDifference

Options for Timestamp::since and Timestamp::until.

TimestampDisplayWithOffset

A type for formatting a Timestamp with a specific offset.

TimestampRound

Options for Timestamp::round.

TimestampSeries

An iterator over periodic timestamps, created by Timestamp::series.

Zoned

A time zone aware instant in time.

ZonedArithmetic

Options for Timestamp::checked_add and Timestamp::checked_sub.

ZonedDifference

Options for Zoned::since and Zoned::until.

ZonedRound

Options for Zoned::round.

ZonedSeries

An iterator over periodic zoned datetimes, created by Zoned::series.

ZonedWith

A builder for setting the fields on a Zoned.

Enums

RoundMode

The mode for dealing with the remainder when rounding datetimes or spans.

Unit

A way to refer to a single calendar or clock unit.

Traits

ToSpan

A trait for enabling concise literals for creating Span values.