pavex::time::fmt::serde

Module timestamp

Convenience routines for (de)serializing Timestamp as raw integer values.

At present, the helpers are limited to serializing and deserializing Timestamp values as an integer number of seconds, milliseconds, microseconds or nanoseconds.

Advice

In general, these helpers should only be used to interface with “legacy” APIs that transmit times as integer number of seconds (or milliseconds or whatever). If you’re designing a new API and need to transmit instants in time that don’t care about time zones, then you should use Timestamp directly. It will automatically use RFC 3339. (And if you do want to include the time zone, then using Zoned directly will work as well by utilizing the RFC 9557 extension to RFC 3339.)

Modules

microsecond

(De)serialize an integer number of microseconds from the Unix epoch.

millisecond

(De)serialize an integer number of milliseconds from the Unix epoch.

nanosecond

(De)serialize an integer number of nanoseconds from the Unix epoch.

second

(De)serialize an integer number of seconds from the Unix epoch.