pavex::time::fmt::serde::unsigned_duration

Module friendly

(De)serialize a std::time::Duration in the friendly duration format.

Note that these will still deserialize ISO 8601 duration strings. The main feature of this module is that serialization will use the friendly format instead of the ISO 8601 format.

Modules

compact

(De)serialize a std::time::Duration in the friendly duration format using compact designators.